requirements.txt
aiofiles==24.1.0
aiogram==3.15.0
asyncpg==0.30.0
envfile==1.0
python-dotenv==1.0.1

